find intersection of 2 lines

+++ Redo/Moderate/LineIntersection/Point.java
@@ -1,15 +1,15 @@
 package Redo.Moderate.LineIntersection;
 
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setPoint(int x, int y) {
+    public void setPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
